+++ internal/admission/server.go
@@ -52,9 +52,9 @@ func (sc *ServerConfig) toTLSConfig(ctx context.Context, log logrus.FieldLogger)
 		if err != nil {
 			return nil, fmt.Errorf("X509KeyPair error: %w", err)
 		}
-		return &tls.Config{ //nolint:gosec
-			MaxVersion:   tls.VersionTLS12,
+		return &tls.Config{
 			MinVersion:   tls.VersionTLS12,
+			MaxVersion:   tls.VersionTLS13,
 			Certificates: []tls.Certificate{keyPair},
 		}, nil
 
@@ -83,9 +83,9 @@ func (sc *ServerConfig) toTLSConfig(ctx context.Context, log logrus.FieldLogger)
 			log.WithError(err).Error("certificate watcher error")
 		}
 	}()
-	return &tls.Config{ //nolint:gosec
-		MaxVersion:     tls.VersionTLS12,
+	return &tls.Config{
 		MinVersion:     tls.VersionTLS12,
+		MaxVersion:     tls.VersionTLS13,
 		GetCertificate: watcher.GetCertificate,
 	}, nil
 }